refactor: cqrs
