+++ src/main/java/mju/iphak/maru_egg/admission/api/AdminAdmissionTypeStatusController.java
@@ -1,4 +1,4 @@
-package mju.iphak.maru_egg.admission.api.admin;
+package mju.iphak.maru_egg.admission.api;
 
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
